Align permission_required decorator behavior to Django 2.1.3
The fix for the Django issue [#28379] arrived in the release 2.1.3,
leading to a smarter implementation of the redirect behavior happening
when an user is logged in but don't have the right permissions.

In the new implementation those users are redirected to the 403 page
instead of to the login URL.

This patch alignes the implementation of the `permission_required`
decorator to the new policy, which django-rules is already respecting
in the `PermissionRequiredMixin` mixin.

See the relative Django commit in [1].

[#28379]: https://code.djangoproject.com/ticket/28379
[1]: https://github.com/django/django/commit/9b1125bfc7e2dc747128e6e7e8a2259ff1a7d39f